feature: implement removing articles from collections

+++ server/src/main/java/com/csc340_group_one/brain_dump/collection/CollectionController.java
@@ -2,6 +2,7 @@ package com.csc340_group_one.brain_dump.collection;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -53,11 +55,27 @@ public class CollectionController {
 
     @PutMapping("{id}")
     public Optional<Collection> updateCollection(@PathVariable Long id, @RequestBody Collection collection) {
+        System.out.println(collection);
         return this.service.updateCollection(id, collection);
     }
 
     @DeleteMapping("{id}")
-    public void deleteCollection(@PathVariable Long id) {
-        this.service.deleteCollection(id);
+    public void deleteCollection(@PathVariable Long id,
+            @RequestParam Optional<Long> articleId) {
+        if (articleId.isEmpty()) {
+            this.service.deleteCollection(id);
+        } else {
+            Optional<Collection> optCollection = this.service.getCollectionById(id);
+            if (optCollection.isEmpty()) {
+                return;
+            }
+            Collection collection = optCollection.get();
+            collection.setArticles(
+                    collection.getArticles()
+                            .stream()
+                            .filter(article -> article.getId() != articleId.get())
+                            .collect(Collectors.toList()));
+            this.service.updateCollection(id, collection);
+        }
     }
 }
